Listo para el CSS. Implementación de MiddleWare, look up de asientos disponibles en DB y Grafica

+++ routes/web.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Pdf\PdfController;
 use App\Http\Controllers\Home\HomeController;
 use App\Http\Controllers\Admin\EventoController;
 use App\Http\Controllers\Boleto\BoletoController;
+use App\Http\Controllers\Grafica\GraficaController;
 
 Route::get('/', function () {
     return view('index');
@@ -18,7 +19,7 @@ Route::middleware([
     'verified',
 ])->group(function () {
     Route::get('/dashboard', function () {
-        return view('dashboard');
+        return view('index');
     })->name('dashboard');
 });
 
@@ -45,9 +46,9 @@ Route::get('/home', function () {
 
 
     
-});
+})->name('home');;
 
-Route::post('/admin-eventos', [EventoController::class, 'store']);
+Route::post('/admin-eventos', [EventoController::class, 'store'])->middleware('admin');
 
 Route::get('/admin-ver_eventos', function(){
     if(Auth::id()){
@@ -62,18 +63,21 @@ Route::get('/admin-ver_eventos', function(){
     }else{
         return redirect()->back();
     }
-});
+})->middleware('admin')->name('admin-ver_eventos');
 
-Route::delete('/admin-eliminar_eventos/{evento}', [EventoController::class, 'eliminarEvento']);
+Route::delete('/admin-eliminar_eventos/{evento}', [EventoController::class, 'eliminarEvento'])->middleware('admin');
 
-Route::get('admin-editar_eventos/{evento}',[EventoController::class, 'mostrarPantallaEdicion']);
-Route::put('admin-editar_eventos/{evento}',[EventoController::class, 'actualizarInfoEvento']);
+Route::get('admin-editar_eventos/{evento}',[EventoController::class, 'mostrarPantallaEdicion'])->middleware('admin')->name('admin.editar_evento');
+Route::put('admin-editar_eventos/{evento}',[EventoController::class, 'actualizarInfoEvento'])->middleware('admin');
 
-Route::get('home-pagina_eventos/{evento}',[HomeController::class, 'mostrarPaginaEvento']);
+Route::get('home-pagina_eventos/{evento}',[HomeController::class, 'mostrarPaginaEvento'])->middleware('user');
 
-Route::post('home-seleccionboleto_eventos/{evento}', [BoletoController::class, 'mostrarPantallaBoletos']);
+Route::post('home-seleccionboleto_eventos/{evento}', [BoletoController::class, 'mostrarPantallaBoletos'])->middleware('user');
 
-Route::post('/comprar-asiento', [BoletoController::class, 'comprarAsiento']);
+Route::post('/comprar-asiento', [BoletoController::class, 'comprarAsiento'])->middleware('user');
 
 //Generar Pdf
-Route::get('/pdf-boleto/{evento}', [PdfController::class, 'generarPdf']);
+Route::get('/pdf-boleto/{evento}', [PdfController::class, 'generarPdf'])->middleware('user');
+
+//Ir a pagina de grafica de eventos - Admin
+Route::get('/admin-grafica_eventos',[GraficaController::class, 'verGrafica'])->middleware('admin');
